recurring weekly events added

+++ calendar/src/main/java/com/homates/calendar/model/Event.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import net.logicsquad.recurring.*;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
@@ -20,7 +19,7 @@ public class Event {
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @Column(name= "description")
     private String description;
